requirements.txt
aiomysql==0.3.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
certifi==2025.10.5
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.120.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
